feat: Sections 일급 콜렉션으로 변경
