+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -18,8 +18,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections;
 
     public Line() {
     }
@@ -32,7 +32,7 @@ public class Line extends BaseEntity {
     public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
-        sections.add(new Section(this, upStation, downStation, distance));
+        this.sections = new Sections(new Section(this, upStation, downStation, distance));
     }
 
     public void update(Line line) {
@@ -52,97 +52,20 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Section> getSections() {
+    public Sections getSections() {
         return sections;
     }
 
     public List<Station> getStations() {
-        Section startSection = findStartSection(sections.get(0));
-        List<Station> stations = new ArrayList<>();
-        stations.add(startSection.getUpStation());
-        Station nextStation = startSection.getDownStation();
-        while (nextStation != null) {
-            stations.add(nextStation);
-            nextStation = findDownStation(nextStation);
-        }
-        return stations;
+        return sections.getStations();
     }
 
     public void addSection(Station upStation, Station downStation, int distance) {
-        List<Station> stations = this.getStations();
-        if (stations.isEmpty()) {
-            sections.add(new Section(this, upStation, downStation, distance));
-            return;
-        }
-        if (stations.contains(upStation) == stations.contains(downStation)) {
-            throw new RuntimeException();
-        }
-        tryAddToUp(upStation, downStation, distance);
-        sections.add(new Section(this, upStation, downStation, distance));
+        sections.addSection(this, upStation, downStation, distance);
     }
 
     public void removeStation(Station station) {
-        checkValidation(station);
-        Optional<Section> upLineStation = sections.stream()
-                .filter(it -> it.getUpStation() == station)
-                .findFirst();
-        Optional<Section> downLineStation = sections.stream()
-                .filter(it -> it.getDownStation() == station)
-                .findFirst();
-        if (upLineStation.isPresent() && downLineStation.isPresent()) {
-            sections.add(createNewSection(upLineStation.get(), downLineStation.get()));
-        }
-        upLineStation.ifPresent(it -> sections.remove(it));
-        downLineStation.ifPresent(it -> sections.remove(it));
-    }
-
-    private void checkValidation(Station station) {
-        if (sections.size() <= 1 || !getStations().contains(station)) {
-            throw new RuntimeException();
-        }
-    }
-
-    private Station findDownStation(Station station) {
-        return this.sections.stream()
-                .filter(it -> it.getUpStation().equals(station))
-                .findFirst()
-                .map(Section::getDownStation)
-                .orElse(null);
-    }
-
-    private Section findStartSection(Section section) {
-        Optional<Section> optionalSection = sections.stream()
-                .filter(it -> it.getDownStation().equals(section.getUpStation()))
-                .findFirst();
-        if (!optionalSection.isPresent()) {
-            return section;
-        }
-        return findStartSection(optionalSection.get());
-    }
-
-    private void tryAddToUp(Station upStation, Station downStation, int distance) {
-        Optional<Section> section = sections.stream()
-                .filter(it -> it.getUpStation() == upStation)
-                .findFirst();
-        if (section.isPresent()) {
-            section.get().updateUpStation(downStation, distance);
-            return;
-        }
-        tryAddToDown(upStation, downStation, distance);
-    }
-
-    private void tryAddToDown(Station upStation, Station downStation, int distance) {
-        Optional<Section> section = sections.stream()
-                .filter(it -> it.getDownStation() == downStation)
-                .findFirst();
-        section.ifPresent(value -> value.updateDownStation(upStation, distance));
-    }
-
-    private Section createNewSection(Section upLineSection, Section downLineSection) {
-        Station newUpStation = downLineSection.getUpStation();
-        Station newDownStation = upLineSection.getDownStation();
-        int newDistance = upLineSection.getDistance() + downLineSection.getDistance();
-        return new Section(this, newUpStation, newDownStation, newDistance);
+        sections.removeStation(this, station);
     }
 
 }
