+++ src/test/java/nextstep/subway/line/LineTest.java
@@ -35,7 +35,7 @@ public class LineTest {
     public void test1() {
         Line line = new Line("2호선", "green", 교대역, 강남역, 10);
         assertThat(line.getStations().size()).isEqualTo(2);
-        assertThat(line.getSections().size()).isEqualTo(1);
+        assertThat(line.getSections().getSections().size()).isEqualTo(1);
         assertThat(line.getStations().get(0)).isEqualTo(교대역);
         assertThat(line.getStations().get(1)).isEqualTo(강남역);
     }
@@ -46,7 +46,7 @@ public class LineTest {
         Line line = new Line("2호선", "green", 교대역, 강남역, 10);
         line.addSection(서초역, 교대역, 10);
         assertThat(line.getStations().size()).isEqualTo(3);
-        assertThat(line.getSections().size()).isEqualTo(2);
+        assertThat(line.getSections().getSections().size()).isEqualTo(2);
         assertThat(line.getStations().get(0)).isEqualTo(서초역);
         assertThat(line.getStations().get(1)).isEqualTo(교대역);
         assertThat(line.getStations().get(2)).isEqualTo(강남역);
@@ -62,7 +62,7 @@ public class LineTest {
         line.addSection(역삼역, 삼성역, 5);
         line.addSection(신림역, 서초역, 5);
         assertThat(line.getStations().size()).isEqualTo(6);
-        assertThat(line.getSections().size()).isEqualTo(5);
+        assertThat(line.getSections().getSections().size()).isEqualTo(5);
         assertThat(line.getStations().get(0)).isEqualTo(신림역);
         assertThat(line.getStations().get(1)).isEqualTo(서초역);
         assertThat(line.getStations().get(2)).isEqualTo(교대역);
